+++ lib/lightgbm.rb
@@ -5,7 +5,6 @@ require "ffi"
 require "lightgbm/utils"
 require "lightgbm/booster"
 require "lightgbm/dataset"
-require "lightgbm/ffi"
 require "lightgbm/version"
 
 # scikit-learn API
@@ -14,6 +13,9 @@ require "lightgbm/regressor"
 
 module LightGBM
   class Error < StandardError; end
+
+  # friendlier error message
+  autoload :FFI, "lightgbm/ffi"
 
   class << self
     def train(params, train_set, num_boost_round: 100, valid_sets: [], valid_names: [], early_stopping_rounds: nil, verbose_eval: true)