+++ lib/lightgbm/ffi.rb
@@ -2,8 +2,12 @@ module LightGBM
   module FFI
     extend ::FFI::Library
     lib_name = "lib_lightgbm.#{::FFI::Platform::LIBSUFFIX}"
-    vendor_lib = File.expand_path("../../vendor/#{lib_name}", __dir__)
-    ffi_lib ["_lightgbm", "lib_lightgbm.so", vendor_lib]
+
+    begin
+      ffi_lib [lib_name, "lib_lightgbm.so"]
+    rescue LoadError => e
+      raise LoadError, "Could not find LightGBM"
+    end
 
     # https://github.com/microsoft/LightGBM/blob/master/include/LightGBM/c_api.h
     # keep same order