+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.4 [unreleased]
 
+- Friendlier message when LightGBM not found
 - Free memory when objects are destroyed
 - Removed unreleased `dump_text` method